+++ CHANGES.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+0.2 (2014-11-30)
+----------------
+
+- Allowing `map_attr` and `store` to be used together.
+- Allowing `Dictionary` containers to be used as attributes.
+- Fixed method resolution order for modified classes.
+- Added a `yorm.settings.fake` option to bypass the filesystem.
+
 0.1.1 (2014-10-20)
 ------------------